requirements.txt
apache-airflow==2.9.3
requests
sqlite-utils
beautifulsoup4

